+++ MANIFEST.in
@@ -4,5 +4,6 @@ recursive-include melissa static css *.css *.otf
 recursive-include melissa static images *.png
 recursive-include melissa static sounds *.js
 recursive-include melissa static *.mp3
+recursive-include melissa utilities pywapi pywapi.py
 include requirements.txt
 include requirements-dev.txt
