+++ setup.py
@@ -50,5 +50,5 @@ setup(
         ],
     },
     zip_safe=False,
-    keywords="virtual assistant speech-to-text text-to-speech melissa jarvis",
+    keywords="virtual-assistant speech-to-text text-to-speech melissa jarvis",
 )
